requirements.txt
imutils==0.5.3
sympy==1.6
numpy==1.18.5
dlib==19.20.0
opencv-python==4.2.0.34
tqdm==4.46.1
torchvision==0.6.1
torch==1.5.1
pillow==7.1.2
adabound==0.0.5
dominate==2.5.1
visdom==0.1.8.8